Add confirmation dialog to user removal on mypage. Also reload the page after removal

+++ massage-booking-system-frontend/src/components/MyPage/MyPage.js
@@ -3,6 +3,8 @@ import { UserContext, NotificationContext } from '../../App'
 import OwnAppointments from '../OwnAppointments/OwnAppointments'
 import useField from '../../hooks/useField'
 import Notification from '../Notification/Notification'
+import { confirmAlert } from 'react-confirm-alert' // Import
+import 'react-confirm-alert/src/react-confirm-alert.css' // Import css
 
 const MyPage = () => {
   const { user, setUser, userService } = useContext(UserContext)
@@ -37,11 +39,24 @@ const MyPage = () => {
   }
 
   const handleRemoveUser = async (id) => {
-    try {
-      await userService.remove(id)
-    } catch (exception) {
-      window.location.reload()
-    }
+    confirmAlert({
+      message: 'Are you sure you want to remove your profile?',
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: async () => {
+            try {
+              await userService.remove(id)
+            } catch (exception) {
+              window.location.reload()
+            }
+          }
+        },
+        {
+          label: 'No',
+        }
+      ]
+    })
   }
 
   // NOTICE -- user && rest rendered. Otherwise nothing gets rendered
